C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik5.cpp -o C:\Users\zainu\Desktop\курс\proekt2\Ozhaev Zaynudin Sharik5.exe  -Wshadow -Winit-self -Wredundant-decls -Wcast-align -Wundef -Wfloat-equal -Winline -Wunreachable-code -Wmissing-declarations -Wmissing-include-dirs -Wswitch-enum -Wswitch-default -Weffc++ -Wmain -Wextra -Wall -g -pipe -fexceptions -Wcast-qual -Wconversion -Wctor-dtor-privacy -Wempty-body -Wformat-security -Wformat=2 -Wignored-qualifiers -Wlogical-op -Wmissing-field-initializers -Wnon-virtual-dtor -Woverloaded-virtual -Wpointer-arith -Wsign-promo -Wstack-usage=8192 -Wstrict-aliasing -Wstrict-null-sentinel -Wtype-limits -Wwrite-strings -D_DEBUG -D_EJUDGE_CLIENT_SIDE -D_EJC

+++ Ozhaev Zaynudin Sharik5.cpp
@@ -223,8 +223,8 @@ viod Bam_balls (int* x, int* y, int* vx, int* vy);
 
 void CollisionBall (Ball* ball_1, Ball* ball_2, int dt)
 {
-    int Dx = (*ball_1) .x - (*ball_2) .x; // стороны треугольника
-    int Dy = (*ball_1) .y - (*ball_2) .y; // стороны треугольника
+    float Dx = (*ball_1) .x - (*ball_2) .x; // стороны треугольника
+    float Dy = (*ball_1) .y - (*ball_2) .y; // стороны треугольника
     double d = sqrt(Dx*Dx + Dy*Dy); if (d == 0) d = 0.01; //гипотенуза
     printf ("Coordin do(): x  = %f and y  = %f\n", (*ball_1) .x, (*ball_2) .x);
     printf ("Coordin do(): Dx  = %f\n", Dx);
